# docs/swagger.yaml
basePath: /
definitions:
...
  example.CreateExampleRequest:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
  example.Example:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      status_id:
        type: integer
      updated_at:
        type: string
    type: object
...
host: localhost:3000
info:
  contact:
    name: API Support
  description: This is a modular REST API built with Go Fiber
  title: My API
  version: "1.0"
paths:
...
  /v1/examples:
    get:
      consumes:
      - application/json
      description: Get list of all examples
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/example.Example'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all examples
      tags:
      - Example
    post:
      consumes:
      - application/json
      description: Create a new example with name and description
      parameters:
      - description: Example data
        in: body
        name: example
        required: true
        schema:
          $ref: '#/definitions/example.CreateExampleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/example.Example'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new example
      tags:
      - Example
  /v1/examples/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete an example (set status_id to 0)
      parameters:
      - description: Example ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Soft delete example
      tags:
      - Example
    get:
      consumes:
      - application/json
      description: Get a specific example by its ID
      parameters:
      - description: Example ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/example.Example'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get example by ID
      tags:
      - Example
    put:
      consumes:
      - application/json
      description: Update an existing example
      parameters:
      - description: Example ID
        in: path
        name: id
        required: true
        type: integer
      - description: Example data
        in: body
        name: example
        required: true
        schema:
          $ref: '#/definitions/example.CreateExampleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/example.Example'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update example
      tags:
      - Example
  /v1/examples/{id}/restore:
    post:
      consumes:
      - application/json
      description: Restore a soft deleted example (set status_id to 1)
      parameters:
      - description: Example ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Restore example
      tags:
      - Example
  /v1/examples/deleted:
    get:
      consumes:
      - application/json
      description: Get list of all soft deleted examples
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/example.Example'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get deleted examples
      tags:
      - Example
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and your API token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"